requirements.txt
pandas==2.2.2
pillow==10.4.0
selenium==4.23.1